Split it in two
